+++ resources/js/components/twc-ui/date-range-picker.tsx
@@ -73,7 +73,7 @@ const DateRangePicker = ({
       isInvalid={hasError}
       value={value}
       onChange={onChange}
-      validationBehavior="native"
+      validationBehavior="aria"
       {...props}
     >
       <Label value={label} />